split pymatviz/io.py out from pymatviz/utils.py
